sha2: Remove common module
