+++ src/hazardous/hash/sha2/mod.rs
@@ -20,6 +20,24 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 // SOFTWARE.
 
+use core::ops::{BitAnd, BitOr, BitXor};
+
+/// The Ch function as specified in FIPS 180-4 section 4.1.3.
+pub(crate) fn ch<T>(x: T, y: T, z: T) -> T
+where
+    T: BitXor<Output = T> + BitAnd<Output = T> + Copy,
+{
+    z ^ (x & (y ^ z))
+}
+
+/// The Maj function as specified in FIPS 180-4 section 4.1.3.
+pub(crate) fn maj<T>(x: T, y: T, z: T) -> T
+where
+    T: BitOr<Output = T> + BitAnd<Output = T> + Copy,
+{
+    (x & y) | (z & (x | y))
+}
+
 macro_rules! func_update (($blocksize:expr, $primitive:ident) => (
     #[must_use = "SECURITY WARNING: Ignoring a Result can have real security implications."]
     /// Update state with `data`. This can be called multiple times.
@@ -155,14 +173,11 @@ macro_rules! func_compress_and_process (($blocksize:expr, $primitive:ident, $def
     }
 ));
 
-/// SHA512 as specified in the [FIPS PUB 180-4](https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.180-4.pdf).
-pub mod sha512;
-
 /// SHA256 as specified in the [FIPS PUB 180-4](https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.180-4.pdf).
 pub mod sha256;
 
 /// SHA384 as specified in the [FIPS PUB 180-4](https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.180-4.pdf).
 pub mod sha384;
 
-/// Common functionality shared between SHA2 variants.
-mod common;
+/// SHA512 as specified in the [FIPS PUB 180-4](https://nvlpubs.nist.gov/nistpubs/FIPS/NIST.FIPS.180-4.pdf).
+pub mod sha512;
